add dao impl and unit test

+++ src/main/java/com/banzhe/hokage/persistence/dao/HokageSecurityGroupDao.java
@@ -1,6 +1,7 @@
 package com.banzhe.hokage.persistence.dao;
 
 import com.banzhe.hokage.persistence.dataobject.HokageSecurityGroupDO;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
@@ -16,14 +17,14 @@ public interface HokageSecurityGroupDao {
      * @param securityGroupDO
      * @return
      */
-    Integer insert(HokageSecurityGroupDO securityGroupDO);
+    Long insert(HokageSecurityGroupDO securityGroupDO);
 
     /**
      * 更新服务器安全组信息
      * @param securityGroupDO
      * @return
      */
-    Integer update(HokageSecurityGroupDO securityGroupDO);
+    Long update(HokageSecurityGroupDO securityGroupDO);
 
     /**
      * 获取所有安全组信息
