+++ src/main/java/com/banzhe/hokage/persistence/dao/HokageTaskDao.java
@@ -16,14 +16,14 @@ public interface HokageTaskDao {
      * @param hokageTaskDO
      * @return
      */
-    Integer insert(HokageTaskDO hokageTaskDO);
+    Long insert(HokageTaskDO hokageTaskDO);
 
     /**
      * 更新一条记录
      * @param hokageTaskDO
      * @return
      */
-    Integer update(HokageTaskDO hokageTaskDO);
+    Long update(HokageTaskDO hokageTaskDO);
 
     /**
      * 根据id查找任务信息
