+++ src/main/java/com/banzhe/hokage/persistence/mapper/HokageSecurityGroupMapper.java
@@ -20,14 +20,14 @@ public interface HokageSecurityGroupMapper {
      * @param securityGroupDO
      * @return
      */
-    Integer insert(HokageSecurityGroupDO securityGroupDO);
+    Long insert(HokageSecurityGroupDO securityGroupDO);
 
     /**
      * 更新服务器安全组信息
      * @param securityGroupDO
      * @return
      */
-    Integer update(HokageSecurityGroupDO securityGroupDO);
+    Long update(HokageSecurityGroupDO securityGroupDO);
 
     /**
      * 获取所有安全组信息
