+++ src/main/java/com/banzhe/hokage/persistence/mapper/HokageServerGroupMapper.java
@@ -20,7 +20,7 @@ public interface HokageServerGroupMapper {
      * @param serverGroupDO
      * @return
      */
-    Integer insert(HokageServerGroupDO serverGroupDO);
+    Long insert(HokageServerGroupDO serverGroupDO);
 
     /**
      * 返回所有服务器组信息
@@ -33,6 +33,6 @@ public interface HokageServerGroupMapper {
      * @param serverGroupDO
      * @return
      */
-    Integer update(HokageServerGroupDO serverGroupDO);
+    Long update(HokageServerGroupDO serverGroupDO);
 
 }
