+++ src/main/java/com/banzhe/hokage/persistence/mapper/HokageTaskResultMapper.java
@@ -20,14 +20,14 @@ public interface HokageTaskResultMapper {
      * @param hokageTaskResultDO
      * @return
      */
-    Integer insert(HokageTaskResultDO hokageTaskResultDO);
+    Long insert(HokageTaskResultDO hokageTaskResultDO);
 
     /**
      * 更新一条任务执行结果
      * @param hokageTaskResultDO
      * @return
      */
-    Integer update(HokageTaskResultDO hokageTaskResultDO);
+    Long update(HokageTaskResultDO hokageTaskResultDO);
 
     /**
      * 根据id主键查找任务执行结果
